Fixing the file structure

+++ testing/integration_tests/mock_test_cbmed_runs.py
@@ -38,11 +38,13 @@ def setup_environment():
 
     if user_input.lower() == 'y':
         print("Proceeding with teardown...")
-
-        rmtree(test_cbmed_run_seq_dir)
+        
+        if test_cbmed_run_seq_dir.exists():
+            rmtree(test_cbmed_run_seq_dir)
         print(f"Removed directory: {test_cbmed_run_seq_dir}")
 
-        rmtree(test_results)
+        if test_results.exists():
+            rmtree(test_results)
         print(f"Removed directory: {test_results}")
 
     else:
